Moved NBitcoin Stratis node tests to Integration Tests Removed pos_RPCClientTests and moved all the tests to the relevant test classes. Removed NodeBuilderStratis from NBitcoin.Tests
