+++ src/NBitcoin.Tests/util_tests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Text;
 using NBitcoin.BouncyCastle.Math;
 using NBitcoin.Crypto;
@@ -891,6 +892,34 @@ namespace NBitcoin.Tests
             var b1_2 = b1 - (b2) + (new Money(10000));
             Assert.True(
                 b1_2.SequenceEqual(new IMoney[] { new AssetMoney(msft, 9), new AssetMoney(goog, 8), new Money(10000) }));
+        }
+
+        [Fact]
+        [Trait("UnitTest", "UnitTest")]
+        public void CanParseIpEndpoint()
+        {
+            IPEndPoint endpoint = Utils.ParseIpEndpoint("google.com:94", 90);
+            Assert.Equal(94, endpoint.Port);
+
+            endpoint = Utils.ParseIpEndpoint("google.com", 90);
+            Assert.Equal(90, endpoint.Port);
+
+            endpoint = Utils.ParseIpEndpoint("10.10.1.3", 90);
+            Assert.Equal("10.10.1.3", endpoint.Address.ToString());
+            Assert.Equal(90, endpoint.Port);
+
+            endpoint = Utils.ParseIpEndpoint("10.10.1.3:94", 90);
+            Assert.Equal("10.10.1.3", endpoint.Address.ToString());
+            Assert.Equal(94, endpoint.Port);
+            Assert.Throws<System.Net.Sockets.SocketException>(() => Utils.ParseIpEndpoint("2001:db8:1f70::999:de8:7648:6e8:100", 90));
+
+            endpoint = Utils.ParseIpEndpoint("2001:db8:1f70::999:de8:7648:6e8", 90);
+            Assert.Equal("2001:db8:1f70:0:999:de8:7648:6e8", endpoint.Address.ToString());
+            Assert.Equal(90, endpoint.Port);
+
+            endpoint = Utils.ParseIpEndpoint("[2001:db8:1f70::999:de8:7648:6e8]:94", 90);
+            Assert.Equal("2001:db8:1f70:0:999:de8:7648:6e8", endpoint.Address.ToString());
+            Assert.Equal(94, endpoint.Port);
         }
     }
 }
