+++ src/Stratis.Bitcoin.IntegrationTests/RPC/RPCTests.cs
@@ -27,7 +27,7 @@ namespace Stratis.Bitcoin.IntegrationTests.RPC
 
             // Move a wallet file to the right folder and restart the wallet manager to take it into account.
             this.InitializeTestWallet(this.Node.FullNode.DataFolder.WalletPath);
-            var walletManager = this.Node.FullNode.NodeService<IWalletManager>() as WalletManager; ;
+            var walletManager = this.Node.FullNode.NodeService<IWalletManager>() as WalletManager;
             walletManager.Start();
         }
     }
@@ -62,15 +62,6 @@ namespace Stratis.Bitcoin.IntegrationTests.RPC
         {
             var hash = this.rpcTestFixture.RpcClient.GetBestBlockHash();
 
-            try
-            {
-                this.rpcTestFixture.RpcClient.SendCommand("lol");
-                Assert.True(false, "should throw");
-            }
-            catch (RPCException ex)
-            {
-                Assert.Equal(RPCErrorCode.RPC_METHOD_NOT_FOUND, ex.RPCCode);
-            }
             Assert.Equal(hash, Network.RegTest.GetGenesis().GetHash());
             var oldClient = this.rpcTestFixture.RpcClient;
             var client = new RPCClient("abc:def", this.rpcTestFixture.RpcClient.Address, this.rpcTestFixture.RpcClient.Network);
@@ -94,6 +85,41 @@ namespace Stratis.Bitcoin.IntegrationTests.RPC
             {
                 Assert.Equal(RPCErrorCode.RPC_MISC_ERROR, ex.RPCCode);
             }
+        }
+
+        [Fact]
+        public void InvalidCommandSendRPCException()
+        {
+            RPCException ex = Assert.Throws<RPCException>(() => this.rpcTestFixture.RpcClient.SendCommand("donotexist"));
+            Assert.True(ex.RPCCode == RPCErrorCode.RPC_METHOD_NOT_FOUND);
+        }
+
+        [Fact]
+        public void CanSendCommand()
+        {
+            RPCResponse response = this.rpcTestFixture.RpcClient.SendCommand(RPCOperations.getinfo);
+            Assert.NotNull(response.Result);
+        }
+
+        [Fact]
+        public void CanGetRawMemPool()
+        {
+            uint256[] ids = this.rpcTestFixture.RpcClient.GetRawMempool();
+            Assert.NotNull(ids);
+        }
+
+        [Fact]
+        public void CanGetBlockCount()
+        {
+            var blockCount = this.rpcTestFixture.RpcClient.GetBlockCountAsync().Result;
+            Assert.Equal(0, blockCount);
+        }
+
+        [Fact]
+        public void CanGetStratisPeersInfo()
+        {
+            var peers = this.rpcTestFixture.RpcClient.GetStratisPeersInfoAsync().Result;
+            Assert.NotEmpty(peers);
         }
 
         /// <summary>
@@ -185,5 +211,130 @@ namespace Stratis.Bitcoin.IntegrationTests.RPC
             string resp = this.rpcTestFixture.RpcClient.SendCommand("generate", "1").ResultString;
             Assert.StartsWith("[" + Environment.NewLine + "  \"", resp);
         }
+
+        // TODO: implement the RPC methods used below
+        //[Fact]
+        //public void RawTransactionIsConformsToRPC()
+        //{
+        //   var tx = Transaction.Load("01000000ac55a957010000000000000000000000000000000000000000000000000000000000000000ffffffff0401320103ffffffff010084d717000000001976a9143ac0dad2ad42e35fcd745d7511d47c24ad6580b588ac00000000", network: this.rpcTestFixture.Node.FullNode.Network);
+        //   Transaction tx2 = this.rpcTestFixture.RpcClient.GetRawTransaction(uint256.Parse("65a26bc20b0351aebf05829daefa8f7db2f800623439f3c114257c91447f1518"));
+        //   Assert.Equal(tx.GetHash(), tx2.GetHash());
+        //}
+
+        //[Fact]
+        //public void CanDecodeRawTransaction()
+        //{
+        //    var tx = this.rpcTestFixture.Node.FullNode.Network.GetGenesis().Transactions[0];
+        //    var tx2 = this.rpcTestFixture.RpcClient.DecodeRawTransaction(tx.ToBytes());
+        //    Assert.True(JToken.DeepEquals(tx.ToString(RawFormat.Satoshi), tx2.ToString(RawFormat.Satoshi)));
+        //}
+
+        //[Fact]
+        //public void CanEstimateFeeRate()
+        //{
+        //    Assert.Throws<NoEstimationException>(() => this.rpcTestFixture.RpcClient.EstimateFeeRate(1));
+        //}
+
+        //[Fact]
+        //public void TryEstimateFeeRate()
+        //{
+        //    Assert.Null(this.rpcTestFixture.RpcClient.TryEstimateFeeRate(1));
+        //}
+
+        //[Fact]
+        //public void CanGetTransactionBlockFromRPC()
+        //{
+        //    uint256 blockId = this.rpcTestFixture.RpcClient.GetBestBlockHash();
+        //    RPCBlock block = this.rpcTestFixture.RpcClient.GetRPCBlockAsync(blockId).Result;
+        //    Assert.NotNull(block);
+        //    Assert.Equal(blockId, uint.Parse(block.hash));
+        //}
+
+        //[Fact]
+        //public void CanAddNodes()
+        //{
+
+        //    using (var builder = NodeBuilder.Create(this))
+        //    {
+        //        CoreNode nodeA = builder.CreateStratisPosNode();
+        //        CoreNode nodeB = builder.CreateStratisPosNode();
+        //        builder.StartAll();
+
+        //        RPCClient rpc = nodeA.CreateRPCClient();
+        //        rpc.RemoveNode(nodeA.Endpoint);
+        //        rpc.AddNode(nodeB.Endpoint);
+        //        Thread.Sleep(500);
+
+        //        AddedNodeInfo[] info = rpc.GetAddedNodeInfo(true);
+        //        Assert.NotNull(info);
+        //        Assert.NotEmpty(info);
+
+        //        //For some reason this one does not pass anymore in 0.13.1.
+        //        //Assert.Equal(nodeB.Endpoint, info.First().Addresses.First().Address);
+        //        AddedNodeInfo oneInfo = rpc.GetAddedNodeInfo(true, nodeB.Endpoint);
+        //        Assert.NotNull(oneInfo);
+        //        Assert.True(oneInfo.AddedNode.ToString() == nodeB.Endpoint.ToString());
+
+        //        oneInfo = rpc.GetAddedNodeInfo(true, nodeA.Endpoint);
+        //        Assert.Null(oneInfo);
+
+        //        //rpc.RemoveNode(nodeB.Endpoint);
+        //        //Thread.Sleep(500);
+        //        //info = rpc.GetAddedNodeInfo(true);
+        //        //Assert.Equal(0, info.Count());
+        //    }
+        //}
+
+        //[Fact]
+        //public void CanGetPrivateKeysFromAccount()
+        //{
+        //    string accountName = "account";
+        //    Key key = new Key();
+        //    this.rpcTestFixture.RpcClient.ImportAddress(key.PubKey.GetAddress(Network.StratisMain), accountName, false);
+        //    BitcoinAddress address = this.rpcTestFixture.RpcClient.GetAccountAddress(accountName);
+        //    BitcoinSecret secret = this.rpcTestFixture.RpcClient.DumpPrivKey(address);
+        //    BitcoinSecret secret2 = this.rpcTestFixture.RpcClient.GetAccountSecret(accountName);
+
+        //    Assert.Equal(secret.ToString(), secret2.ToString());
+        //    Assert.Equal(address.ToString(), secret.GetAddress().ToString());
+        //}
+
+        //[Fact]
+        //public void CanGetPrivateKeysFromAccount()
+        //{
+        //    string accountName = "account";
+        //    Key key = new Key();
+        //    this.rpcTestFixture.RpcClient.ImportAddress(key.PubKey.GetAddress(Network.StratisMain), accountName, false);
+        //    BitcoinAddress address = this.rpcTestFixture.RpcClient.GetAccountAddress(accountName);
+        //    BitcoinSecret secret = this.rpcTestFixture.RpcClient.DumpPrivKey(address);
+        //    BitcoinSecret secret2 = this.rpcTestFixture.RpcClient.GetAccountSecret(accountName);
+
+        //    Assert.Equal(secret.ToString(), secret2.ToString());
+        //    Assert.Equal(address.ToString(), secret.GetAddress().ToString());
+        //}
+
+        //[Fact]
+        //public void CanBackupWallet()
+        //{
+        //    var buildOutputDir = Path.GetDirectoryName(".");
+        //    var filePath = Path.Combine(buildOutputDir, "wallet_backup.dat");
+        //    try
+        //    {
+        //        this.rpcTestFixture.RpcClient.BackupWallet(filePath);
+        //        Assert.True(File.Exists(filePath));
+        //    }
+        //    finally
+        //    {
+        //        if (File.Exists(filePath))
+        //            File.Delete(filePath);
+        //    }
+        //}
+
+        //[Fact]
+        //public void CanEstimatePriority()
+        //{
+        //    var priority = this.rpcTestFixture.RpcClient.EstimatePriority(10);
+        //    Assert.True(priority > 0 || priority == -1);
+        //}
     }
 }
